pyproject.toml
[build-system]
requires = ["setuptools>=42.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.pyright]
include = ["bot", "handlers", "utils", "config"]
exclude = ["**/__pycache__"]
reportMissingImports = false
reportMissingModuleSource = false

[tool.black]
line-length = 88
target-version = ['py38']

[project]
name = "location-tracker-bot"
version = "1.0.0"
description = "Telegram bot for tracking locations"
requires-python = ">=3.8"
dependencies = [
    "python-telegram-bot>=20.0",
    "requests>=2.25.1",
    "python-dateutil>=2.8.2",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "black",
    "mypy",
    "pytest",
] 